code/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
python-dotenv==1.0.0
chromadb==0.4.22
tiktoken==0.5.2
faiss-cpu==1.7.4
